# Avoid database lookups for `DescriptionHTML`