- Don't Compose meta's for DescriptionHTML, they are only needed in order to correctly format and show issue's but it's highly unlikely that a repository description will refer to a local issue.
- Improve speed that `/explore/repos` can be rendered(this template is the most noticeable in performance improvements, as it renders `DescriptionHTML` for ~20 repos).
- `ComposeMetas` lookups for ExternalTracker Unit and if the repo was from a organisation it would do another lookup for team names + repo owner.

Using 125 Connections for 5 seconds: on `/explore/repos`:

Before:
```
Statistics        Avg      Stdev        Max
  Reqs/sec       569.41     506.05    2715.00
  Latency      214.27ms    16.60ms   294.84ms
  HTTP codes:
    1xx - 0, 2xx - 2974, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    27.17MB/s
```
After:
```
Statistics        Avg      Stdev        Max
  Reqs/sec      1585.04     789.84    4144.56
  Latency       78.89ms    15.89ms   206.94ms
  HTTP codes:
    1xx - 0, 2xx - 7975, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    73.85MB/s
```